Split puppet.rs into puppet/{mod.rs,display.rs}
